Step 1, put each component into its own file
